Portability is the name of the game. No longer assuming long's are 32bits (not true for 64bit GCC)

+++ lib/vpc.h
@@ -10,6 +10,8 @@
     date:       February 13th, 2015
 */
 
+#include <stdint.h>
+
 /* 16kB = (16 (0x10) * 1024B (0x400)) */
 #define TOTAL_MEMORY    0x4000
 
@@ -22,13 +24,13 @@
 void writef(void *memory);
 
 /* Load a file into the memory. */
-int loadf(void *memory, unsigned int max);
+int loadf(void *memory, uint16_t max);
 
 /* Dumps the memory to screen. */
-void dumpmem(void *memory, unsigned int offset, unsigned int length);
+void dumpmem(void *memory, uint16_t offset, uint16_t length);
 
 /* Modifies the memory. */
-void modmem(void *memory, unsigned int offset);
+void modmem(void *memory, uint16_t offset);
 
 
 /**********/
